"Improve" vite "alias"
Guess user intent to distinguish if its approrpriate to add flags to
vite. Vite without any specified subcommands will start a development
web server, hosting your javascript thing. But it defaults to only
listen on localhost, and as we run it in a container, this means that we
can't access it. We therefore replace it with a --host flag with a value
of 0.0.0.0 (listen on any address/interface). This however, only works
for this "null" subcommand, but other subcommands, such as build, does
not accept the --host flag.

This is basically tested code... I mean, I wrote it, isn't that enough?
(I did test it as some point, but now I tweaked it and tidied it up a
bit before commiting, so have fun troubleshooting!)